bdd dump question

+++ wpgitinit.sh
@@ -9,12 +9,12 @@ then
   exit $E_USAGE
 fi
 
-EXPECTED_ARGS=4
+EXPECTED_ARGS=3
 E_BADARGS=65
 
 if [ $# -ne $EXPECTED_ARGS ]
 then
-  echo "Usage: bash wpgitinit.sh httpsgitavecmdp gituseremail gitusername bddname"
+  echo "Usage: bash wpgitinit.sh httpsgitavecmdp gituseremail gitusername"
   exit $E_BADARGS
 fi
 
@@ -24,11 +24,16 @@ die() {
         exit 1 ;
 }
 
-rm -rf ./sqldump
-mkdir ./sqldump
+read -p "Dump de la base ? <y/N> " prompt
+if [[ $prompt == "y" || $prompt == "Y" || $prompt == "yes" || $prompt == "Yes" ]]
+then
+	read -p "Nom de la base : " nombdd
+	rm -rf ./sqldump
+	mkdir ./sqldump
 
-mysqldump -uroot -p --databases $4 | /bin/gzip -9 > ./sqldump/$4.sql.gz
-[ $? -eq 0 ] || die "Impossible de faire le dump la base, mot de passe mysql incorrect ?" ;
+	mysqldump -uroot -p --databases $nombdd | /bin/gzip -9 > ./sqldump/$nombdd.sql.gz
+	[ $? -eq 0 ] || die "Impossible de faire le dump la base, mot de passe mysql incorrect ?" ;	
+fi
 
 # Création du repo local git et init
 git init
@@ -39,7 +44,7 @@ touch .gitignore
 echo 'htdocs/wp-content/cache/*' > .gitignore
 echo 'htdocs/wp-content/uploads/backwpup*' >> .gitignore
 git add .
-git commit -m "commit init"
+git commit -a -m "commit init"
 git push origin master
 
 echo "================================================================="
